+++ app/src/main/java/com/example/yp_playlist_maker/data/impl/PlayTrackRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package com.example.yp_playlist_maker.data.impl
+
+class PlayTrackRepositoryImpl {
+}
