+++ app/src/main/java/com/example/yp_playlist_maker/domain/api/repository/PlayTrackRepository.kt
@@ -1,0 +1,4 @@
+package com.example.yp_playlist_maker.domain.api.repository
+
+interface PlayTrackRepository {
+}
