--- app/src/main/java/com/example/yp_playlist_maker/domain/use_case/PlayTrackUseCase.kt
@@ -1,27 +1,0 @@
-package com.example.yp_playlist_maker.domain.use_case
-
-import com.example.yp_playlist_maker.domain.api.interactor.PlayTrackInteractor
-
-class PlayTrackUseCase(private val interactor: PlayTrackInteractor) {
-
-    fun pause() {
-        interactor.pausePlayer()
-    }
-
-    fun preparePlayer() {
-        interactor.preparePlayer()
-    }
-
-    fun playBackControl() {
-        interactor.playbackControl()
-    }
-
-    fun releasePlayer() {
-        interactor.releasePlayer()
-    }
-
-    fun threadRemoveCallbacks() {
-        interactor.threadRemoveCallbacks()
-    }
-
-}
